Split main startup responsibilities into submodules
Decompose src/main.rs into focused CLI and startup modules to address\nCodeScene low-cohesion findings without changing runtime behaviour.\n\nKeep the binary entry point thin, move startup orchestration into\nstartup submodules, and update repository documentation to reflect the\nnew layout.\n\nUse main_cli.rs instead of cli.rs because the crate already owns\nsrc/cli/ as a library module path, and Rust would treat a sibling\nsrc/cli.rs as a conflicting module definition.
