+++ src/startup/mod.rs
@@ -1,0 +1,7 @@
+//! Binary startup helpers for phased bootstrap and channel wiring.
+
+pub(crate) mod boot;
+pub(crate) mod channels;
+pub(crate) mod phases;
+pub(crate) mod run;
+pub(crate) mod wasm;
